게시물 수정, 삭제 시 인증 체크 및 삭제 권한 판별 - Spring Security가 아닌 세션을 활용해 인증-인가 여부 체크

+++ src/main/java/com/brad/community/controller/UsrArticleController.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Controller
@@ -19,11 +20,17 @@ public class UsrArticleController {
 
     @RequestMapping("/doAdd")
     @ResponseBody
-    public DataResponse<Article> doAdd(String title, String body) {
+    public DataResponse<Article> doAdd(HttpSession session, String title, String body) {
+        boolean isLogin = false;
+        if(session.getAttribute("loginMemberId") == null) {
+            return DataResponse.of("F-Authentication", "로그인 후 이용해주세요.");
+        }
+        Long memberId = (Long) session.getAttribute("loginMemberId");
+
         if (Ut.isEmpty(title)) return DataResponse.of("F-1", "게시물 제목을 입력해주세요.");
         if (Ut.isEmpty(body)) return DataResponse.of("F-1", "게시물 내용을 입력해주세요.");
 
-        Long articleId = articleService.writeArticle(title, body);
+        Long articleId = articleService.writeArticle(memberId, title, body);
         Article article = articleService.getArticle(articleId);
         return DataResponse.of("S-1", Ut.f("%d번 게시물이 생성되었습니다.", articleId), article);
     }
@@ -47,22 +54,41 @@ public class UsrArticleController {
 
     @RequestMapping("/doDelete")
     @ResponseBody
-    public DataResponse doDelete(Long id) {
+    public DataResponse doDelete(HttpSession session, Long id) {
+
+        if(session.getAttribute("loginMemberId") == null) {
+            return DataResponse.of("F-Authentication", "로그인 후 이용해주세요.");
+        }
+        Long memberId = (Long) session.getAttribute("loginMemberId");
+
         Article article = articleService.getArticle(id);
         if(article == null) {
             return DataResponse.of("F-1", Ut.f("%d번 게시물이 존재하지 않습니다!", id));
         }
+
+        boolean canDelete = articleService.canDelete(memberId, article);
+        if(!canDelete) return DataResponse.of("F-Authorization", "삭제 권한이 없습니다!");
+
         articleService.deleteArticle(id);
         return DataResponse.of("S-1", Ut.f("%d번 게시물이 삭제되었습니다.", id));
     }
 
     @RequestMapping("/doModify")
     @ResponseBody
-    public DataResponse doModify(Long id, String title, String body) {
+    public DataResponse doModify(HttpSession session, Long id, String title, String body) {
+        if(session.getAttribute("loginMemberId") == null) {
+            return DataResponse.of("F-Authentication", "로그인 후 이용해주세요.");
+        }
+        Long memberId = (Long) session.getAttribute("loginMemberId");
+
         Article article = articleService.getArticle(id);
         if(article == null) {
             return DataResponse.of("F-1", Ut.f("%d번 게시물이 존재하지 않습니다!", id));
         }
+
+        boolean canModify = articleService.canModify(memberId, article);
+        if(!canModify) return DataResponse.of("F-Authorization", "수정 권한이 없습니다!");
+
         articleService.modifyArticle(id, title, body);
         return DataResponse.of("S-1", Ut.f("%d번 게시물이 수정되었습니다.", id));
     }
